Also clean up Visual Studio build warning in wasm::toSInteger32 for double->int32_t cast.

+++ src/support/safe_integer.cpp
@@ -44,7 +44,7 @@ int32_t wasm::toSInteger32(double x) {
   assert(isSInteger32(x));
   return x > std::numeric_limits<int32_t>::min() &&
                  x < std::numeric_limits<int32_t>::max()
-             ? x
+             ? (int32_t)x
              : (x < 0 ? std::numeric_limits<int32_t>::min()
                       : std::numeric_limits<int32_t>::max());
 }
